requirements.txt
paho-mqtt==1.6.1
flask==2.3.3
flask-cors==4.0.0
schedule==1.2.0
